+++ packages/amy-synth-m5/include/AmySynthSlot.h
@@ -10,10 +10,10 @@
 class AmySynthSlot {
  public:
   // Configures the AMY synth slot and immediately applies the initial patch.
-  void begin(uint8_t synthId, uint8_t voiceCount, uint8_t initialPatch);
+  void begin(uint8_t synthId, uint8_t voiceCount, uint16_t initialPatch);
 
   // Selects one of AMY's patch numbers for this synth slot.
-  void setPatch(uint8_t patchNumber);
+  void setPatch(uint16_t patchNumber);
 
   // Sends a note-on event for this synth slot.
   void noteOn(uint8_t midiNote, float velocity);
@@ -23,11 +23,11 @@ class AmySynthSlot {
 
   uint8_t synthId() const;
   uint8_t voiceCount() const;
-  uint8_t patchNumber() const;
+  uint16_t patchNumber() const;
 
  private:
   uint8_t synthId_ = 0;
   uint8_t voiceCount_ = 1;
-  uint8_t patchNumber_ = 0;
+  uint16_t patchNumber_ = 0;
   bool begun_ = false;
 };
